hmm_detection: move rule category definitions/helpers to common
